Refactor render pipeline
